Clean up newlines on descriptions for kuark production and half-life
The line separators in the part descriptions are sort of a hodgepodge of
some in the localization file and some in the code, which led to a
missing newline before the half-life line in the WT-SIT cyclortron
part's description.  Since that newline before the half-life is added by
code in the StationExperiment class, and it seems more appropriate to
have that sort of thing in code rather than the localization file, I've
added it to the code in the SampleAnalyzer class too.  For consistency,
I've also changed the kuark production line (which follows the
half-life) to have its newline added in code as well, rather than
relying on it being at the end of the half-life localized string.